+++ core/api/typescript/src/kurtosis_core_rpc_api_bindings/api_container_service_grpc_pb.js
@@ -137,6 +137,17 @@ function deserialize_api_container_api_LoadModuleResponse(buffer_arg) {
   return api_container_service_pb.LoadModuleResponse.deserializeBinary(new Uint8Array(buffer_arg));
 }
 
+function serialize_api_container_api_PauseServiceArgs(arg) {
+  if (!(arg instanceof api_container_service_pb.PauseServiceArgs)) {
+    throw new Error('Expected argument of type api_container_api.PauseServiceArgs');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_PauseServiceArgs(buffer_arg) {
+  return api_container_service_pb.PauseServiceArgs.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
 function serialize_api_container_api_RegisterServiceArgs(arg) {
   if (!(arg instanceof api_container_service_pb.RegisterServiceArgs)) {
     throw new Error('Expected argument of type api_container_api.RegisterServiceArgs');
@@ -256,6 +267,17 @@ function serialize_api_container_api_UnloadModuleArgs(arg) {
 
 function deserialize_api_container_api_UnloadModuleArgs(buffer_arg) {
   return api_container_service_pb.UnloadModuleArgs.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
+function serialize_api_container_api_UnpauseServiceArgs(arg) {
+  if (!(arg instanceof api_container_service_pb.UnpauseServiceArgs)) {
+    throw new Error('Expected argument of type api_container_api.UnpauseServiceArgs');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_UnpauseServiceArgs(buffer_arg) {
+  return api_container_service_pb.UnpauseServiceArgs.deserializeBinary(new Uint8Array(buffer_arg));
 }
 
 function serialize_api_container_api_UploadFilesArtifactArgs(arg) {
@@ -434,6 +456,30 @@ execCommand: {
     requestDeserialize: deserialize_api_container_api_ExecCommandArgs,
     responseSerialize: serialize_api_container_api_ExecCommandResponse,
     responseDeserialize: deserialize_api_container_api_ExecCommandResponse,
+  },
+  // Pauses all processes running in the service container
+pauseService: {
+    path: '/api_container_api.ApiContainerService/PauseService',
+    requestStream: false,
+    responseStream: false,
+    requestType: api_container_service_pb.PauseServiceArgs,
+    responseType: google_protobuf_empty_pb.Empty,
+    requestSerialize: serialize_api_container_api_PauseServiceArgs,
+    requestDeserialize: deserialize_api_container_api_PauseServiceArgs,
+    responseSerialize: serialize_google_protobuf_Empty,
+    responseDeserialize: deserialize_google_protobuf_Empty,
+  },
+  // Unpauses all paused processes running in the service container
+unpauseService: {
+    path: '/api_container_api.ApiContainerService/UnpauseService',
+    requestStream: false,
+    responseStream: false,
+    requestType: api_container_service_pb.UnpauseServiceArgs,
+    responseType: google_protobuf_empty_pb.Empty,
+    requestSerialize: serialize_api_container_api_UnpauseServiceArgs,
+    requestDeserialize: deserialize_api_container_api_UnpauseServiceArgs,
+    responseSerialize: serialize_google_protobuf_Empty,
+    responseDeserialize: deserialize_google_protobuf_Empty,
   },
   // Block until the given HTTP endpoint returns available, calling it through a HTTP Get request
 waitForHttpGetEndpointAvailability: {