+++ core/api/typescript/src/lib/enclaves/grpc_node_api_container_client.ts
@@ -22,6 +22,8 @@ import {
     ExecuteModuleArgs,
     ExecCommandArgs,
     ExecCommandResponse,
+    PauseServiceArgs,
+    UnpauseServiceArgs,
     UploadFilesArtifactArgs,
     UploadFilesArtifactResponse,
     StoreWebFilesArtifactArgs,
@@ -32,6 +34,7 @@ import {
 import type { ApiContainerServiceClient as ApiContainerServiceClientNode } from "../../kurtosis_core_rpc_api_bindings/api_container_service_grpc_pb";
 import { GenericApiContainerClient } from "./generic_api_container_client";
 import { EnclaveID } from "./enclave_context";
+import {Empty} from "google-protobuf/google/protobuf/empty_pb";
 
 export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
 
@@ -345,6 +348,44 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
 
         const execCommandResponse = execCommandResponseResult.value;
         return ok(execCommandResponse)
+    }
+
+
+    public async pauseService(pauseServiceArgs: PauseServiceArgs): Promise<Result<null, Error>> {
+        const pauseServicePromise: Promise<Result<null, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.pauseService(pauseServiceArgs, (error: ServiceError | null) => {
+                if (error === null) {
+                    resolve(ok(null))
+                } else {
+                    resolve(err(error));
+                }
+            })
+        });
+        const pauseServiceResult: Result<null, Error> = await pauseServicePromise;
+        if(pauseServiceResult.isErr()){
+            return err(pauseServiceResult.error)
+        }
+
+        return ok(null)
+    }
+
+
+    public async unpauseService(unpauseServiceArgs: UnpauseServiceArgs): Promise<Result<null, Error>> {
+        const unpauseServicePromise: Promise<Result<null, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.unpauseService(unpauseServiceArgs, (error: ServiceError | null) => {
+                if (error === null) {
+                    resolve(ok(null))
+                } else {
+                    resolve(err(error));
+                }
+            })
+        });
+        const unpauseServiceResult: Result<null, Error> = await unpauseServicePromise;
+        if(unpauseServiceResult.isErr()){
+            return err(unpauseServiceResult.error)
+        }
+
+        return ok(null)
     }
 
     public async uploadFiles(uploadFilesArtifactArgs: UploadFilesArtifactArgs): Promise<Result<UploadFilesArtifactResponse, Error>> {